Break components out with router to load a tab by URL path
